pyproject.toml
[project]
name = "aps-viewer-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.5",
]

[dependency-groups]
test = [
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aps_viewer_sdk"]